+++ spec/features/staff/users_can_view_budgets_on_an_activity_page_spec.rb
@@ -45,7 +45,7 @@ RSpec.feature "Users can view budgets on an activity page" do
         visit organisation_path(user.organisation)
 
         click_link fund_activity.title
-        click_on I18n.t("tabs.activity.details")
+        click_on I18n.t("tabs.activity.children")
         click_link programme_activity.title
 
         budget_information_is_shown_on_page(budget_presenter)
@@ -65,9 +65,9 @@ RSpec.feature "Users can view budgets on an activity page" do
         visit organisation_path(user.organisation)
 
         click_link fund_activity.title
-        click_on I18n.t("tabs.activity.details")
+        click_on I18n.t("tabs.activity.children")
         click_link programme_activity.title
-        click_on I18n.t("tabs.activity.details")
+        click_on I18n.t("tabs.activity.children")
         click_link project_activity.title
 
         budget_information_is_shown_on_page(budget_presenter)
@@ -83,9 +83,9 @@ RSpec.feature "Users can view budgets on an activity page" do
         visit organisation_path(user.organisation)
 
         click_link fund_activity.title
-        click_on I18n.t("tabs.activity.details")
+        click_on I18n.t("tabs.activity.children")
         click_link programme_activity.title
-        click_on I18n.t("tabs.activity.details")
+        click_on I18n.t("tabs.activity.children")
         click_link project_activity.title
 
         expect(page).to_not have_content(I18n.t("page_content.budgets.button.create"))
@@ -139,7 +139,7 @@ RSpec.feature "Users can view budgets on an activity page" do
         visit organisation_path(user.organisation)
 
         click_link programme_activity.title
-        click_on I18n.t("tabs.activity.details")
+        click_on I18n.t("tabs.activity.children")
         click_link project_activity.title
 
         budget_information_is_shown_on_page(budget_presenter)
@@ -154,7 +154,7 @@ RSpec.feature "Users can view budgets on an activity page" do
         visit organisation_path(user.organisation)
 
         click_link programme_activity.title
-        click_on I18n.t("tabs.activity.details")
+        click_on I18n.t("tabs.activity.children")
         click_link project_activity.title
 
         expect(page).to have_content(I18n.t("page_content.budgets.button.create"))