+++ spec/helpers/activity_helper_spec.rb
@@ -61,5 +61,46 @@ RSpec.describe ActivityHelper, type: :helper do
         end
       end
     end
+
+    context "when the activity is a fund" do
+      context "and is at the indetifer step i.e. the parent step has been skipped" do
+        it "returns true" do
+          activity = build(:activity, :level_form_state, level: :fund, parent: nil)
+
+          expect(helper.step_is_complete_or_next?(activity: activity, step: :identifier)).to be(true)
+        end
+      end
+    end
+  end
+
+  describe "#link_to_activity_parent" do
+    context "when there is no parent" do
+      it "returns nil" do
+        expect(helper.link_to_activity_parent(parent: nil, user: nil)).to be_nil
+      end
+    end
+
+    context "when the parent is a fund" do
+      context "and the user is delivery partner" do
+        it "returns the parent title without a link" do
+          parent_activity = create(:fund_activity)
+          _activity = create(:programme_activity, parent: parent_activity)
+          user = create(:delivery_partner_user)
+
+          expect(helper.link_to_activity_parent(parent: parent_activity, user: user)).to eql parent_activity.title
+        end
+      end
+    end
+
+    context "when there is a parent" do
+      it "returns a link to the parent" do
+        parent_activity = create(:fund_activity)
+        _activity = create(:programme_activity, parent: parent_activity)
+        user = create(:beis_user)
+
+        expect(helper.link_to_activity_parent(parent: parent_activity, user: user)).to include organisation_activity_path(parent_activity.organisation, parent_activity)
+        expect(helper.link_to_activity_parent(parent: parent_activity, user: user)).to include parent_activity.title
+      end
+    end
   end
 end